refactor(headless): extract git safety/rollback wrapper into prompt-candidate-git.ts (#1409)
Move the self-contained git subsystem out of prompt-candidate-loop.ts into
a new sibling prompt-candidate-git.ts: the PromptCandidateGit interface,
the createCliPromptCandidateGit factory (covering rollbackCommit, the
worktree/index dirty checks, and the gitStatusSnapshot/fingerprint
hashing), assertOnlySystemPromptChanged, assertRegularSystemPromptFile,
and the private git/path helpers only it uses (assertGitHeadUnchanged,
gitHeadSha, isGitTracked, hasGitDiff, gitStatusFiles, gitStatusSnapshot,
gitStatusFingerprint, fileFingerprint, gitDiffFingerprint, statusPath,
findGitRoot, toGitRelativePath).

isPathInside and normalizeGitPath are also exported, since the retained
controller-artifact checks in prompt-candidate-loop.ts still use them.
The trivial pure guards isRecord/isNotFound stay as private copies in
each file, matching the existing codebase convention (no shared utils
module; duplicated across many files).

prompt-candidate-loop.ts re-exports createCliPromptCandidateGit,
CreateCliPromptCandidateGitInput, and PromptCandidateGit to preserve the
public API, drops the now-unused node:child_process / realpathSync /
lstat / isAbsolute / promisify imports, and imports the helpers it still
needs from the new module. prompt-optimization-run.ts imports
createCliPromptCandidateGit directly from the new module. No behavior
change; existing CLI git adapter tests pass unchanged.
